refactor: tablegroup 패키지 분리
