+++ src/main/java/kitchenpos/application/TableGroupUngroupedEventHandler.java
@@ -5,9 +5,9 @@ import static java.util.stream.Collectors.toList;
 import java.util.List;
 import kitchenpos.dao.OrderRepository;
 import kitchenpos.domain.Order;
-import kitchenpos.domain.TableGroupUngroupedEvent;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.OrderTableRepository;
+import kitchenpos.domain.tablegroup.TableGroupUngroupedEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
