+++ src/main/java/kitchenpos/dto/tablegroup/TableGroupResponse.java
@@ -1,10 +1,10 @@
-package kitchenpos.dto;
+package kitchenpos.dto.tablegroup;
 
 import static java.util.stream.Collectors.toList;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 import kitchenpos.dto.table.OrderTableResponse;
 
 public class TableGroupResponse {
