+++ src/main/java/kitchenpos/ui/tablegroup/TableGroupRestController.java
@@ -1,9 +1,9 @@
-package kitchenpos.ui;
+package kitchenpos.ui.tablegroup;
 
 import java.net.URI;
-import kitchenpos.application.TableGroupService;
-import kitchenpos.dto.TableGroupRequest;
-import kitchenpos.dto.TableGroupResponse;
+import kitchenpos.application.tablegroup.TableGroupService;
+import kitchenpos.dto.tablegroup.TableGroupRequest;
+import kitchenpos.dto.tablegroup.TableGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
