+++ src/test/java/kitchenpos/fixture/OrderTableFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.TableGroup;
 import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.tablegroup.TableGroup;
 
 public class OrderTableFixture {
