+++ src/test/java/kitchenpos/fixture/TableGroupFixture.java
@@ -6,10 +6,10 @@ import static java.util.stream.Collectors.toList;
 import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.List;
-import kitchenpos.domain.TableGroup;
 import kitchenpos.domain.table.OrderTable;
-import kitchenpos.dto.TableGroupRequest;
-import kitchenpos.dto.TableGroupRequest.OrderTableIdRequest;
+import kitchenpos.domain.tablegroup.TableGroup;
+import kitchenpos.dto.tablegroup.TableGroupRequest;
+import kitchenpos.dto.tablegroup.TableGroupRequest.OrderTableIdRequest;
 
 public class TableGroupFixture {
